Updated the build scripts to use the latest versions of node modules, cleaned up scripts, updated docs to include upcoming features for BaseWeb and docs.
