--- jakefile.js
@@ -1,3 +1,0 @@
-jake.addListener('complete', function () {
-  process.exit();
-});
