+++ jakelib/watch.jake
@@ -1,18 +1,19 @@
-var u = require('./modules/utility');
-var getjsonsync = require('./modules/getjsonsync');
+var u = require('./modules/node-utility');
 var watch = require('node-watch');
 
-/**
- * The Watch Task
- */
 desc('Watch for change to files and rebuild if they change');
 task('watch', {async: true}, function() {
 
-  u.print_notice('Files are now being watched (ctrl+c to exit)', 'yellow');
+  u.printNotice('Files are now being watched (ctrl+c to exit)', 'yellow');
   
-  var settings = getjsonsync('settings.json');
+  var settings = u.getjsonsync('settings.json');
   
   watch('src/scss/', function(filename) {
+    u.print(filename + ' was changed:', 'cyan');
+    jake.Task['build:scss'].execute();
+  });
+  
+  watch('docs/scss/', function(filename) {
     u.print(filename + ' was changed:', 'cyan');
     jake.Task['build:scss'].execute();
   });
@@ -22,11 +23,6 @@ task('watch', {async: true}, function() {
       u.print(filename + ' was changed:', 'cyan');
       jake.Task['build:js'].execute();
     }
-  });
-  
-  watch('docs/scss/', function(filename) {
-    u.print(filename + ' was changed:', 'cyan');
-    jake.Task['build:scss'].execute();
   });
   
   watch(['docs/templates/layouts/', 'docs/templates/pages/', 'docs/templates/partials/'], function(filename) {
