+++ package.json
@@ -10,21 +10,21 @@
   "year": "2015",
   
   "repository": {
-    "type": "git",
-    "url": "https://github.com/sebnitu/BaseWeb.git"
+    "type" : "git",
+    "url"  : "https://github.com/sebnitu/BaseWeb.git"
   },
   
   "licenses": [{
-    "type": "The MIT License (MIT)",
-    "url": "http://en.wikipedia.org/wiki/MIT_License"
+    "type" : "The MIT License (MIT)",
+    "url"  : "http://en.wikipedia.org/wiki/MIT_License"
   }],
   
   "devDependencies": {
-    "jake": "8.0.10",
-    "mustache": "1.0.0",
-    "node-sass": "2.0.1",
-    "uglify-js": "2.4.16",
-    "imagemin": "3.1.0",
-    "node-watch": "0.3.x"
+    "jake"       : "8.0.12",
+    "node-watch" : "0.3.4",
+    "node-sass"  : "2.1.1",
+    "uglify-js"  : "2.4.19",
+    "mustache"   : "2.0.0",
+    "imagemin"   : "3.1.0"
   }
 }
